requirements.txt
cffi==0.9.2
cryptography==0.8.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.7
django-activity-stream==0.5.1
django-allauth==0.19.1
django-angular==0.7.13
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.1
django-haystack==2.3.1
django-model-utils==2.2
django-postgrespool==0.3.0
django-reversion==1.8.5
django-suit==0.2.12
django-tastypie==0.12.1
django-toolbelt==0.0.1
enum34==1.0.4
gnureadline==6.3.3
gunicorn==19.3.0
ipython==3.0.0
oauthlib==0.7.2
psycopg2==2.6
pyasn1==0.1.7
pycparser==2.10
pyOpenSSL==0.14
python-dateutil==2.4.1
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
SQLAlchemy==0.9.9
sqlparse==0.1.14
static3==0.5.1
Werkzeug==0.10.4
whitenoise==1.0.6
Whoosh==2.6.0
